requirements.txt
flask
mysql-connector-python
python-dotenv
requests
google-auth 
google-auth-oauthlib 
google-auth-httplib2
google-api-python-client
